add reqs file and make integration mod
