+++ requirements.txt
@@ -1,0 +1,3 @@
+numpy==2.3.3
+openpyxl==3.1.5
+pandas==2.3.3
